requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiohttp-socks==0.6.0
aioredis==1.3.1
aiorpcX==0.22.1
aiosignal==1.3.1
amqp==5.1.1
anyio==3.3.0
appdirs==1.4.4
APScheduler==3.6.3
arrow==1.1.1
asgiref==3.4.1
asn1crypto==1.4.0
aspy.yaml==1.3.0
asttokens==2.2.1
async-timeout==4.0.2
atomicwrites==1.4.0
attrdict==2.0.1
attrs==22.2.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.11.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1;python_version<"3.9"
base58==2.1.0
base58check==1.0.2
bcrypt==3.2.0
beautifulsoup4==4.9.3
bech32==1.2.0
behave==1.2.6
behave-django==1.4.0
billiard==3.6.4.0
bit==0.7.2
bitarray==1.2.2
bitcoin==1.1.42
Brotli==1.0.9
cached-property==1.5.2
cachetools==4.2.2
cashaddress==1.0.6
cchardet==2.1.7
celery==5.2.7
certifi==2022.12.7
cffi>=1.14.6
cfgv==3.3.1
channels==3.0.4
channels-redis==3.3.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coinaddr==1.0.1
coincurve>=15.0.1
colorama==0.4.5
constantly==15.1.0
contextvars==2.4
coreapi==2.3.3
coreschema==0.0.4
cosmospy==6.0.0
coverage==5.5
crypto==1.4.1
cryptography==39.0.1
cryptos~=2.0.6
cssselect2==0.7.0
cytoolz==0.11.0
daphne==3.0.2
dateparser==1.1.7
decorator==5.0.9
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
dj-rest-auth==2.2.6
Django==3.2.18
django-admin-rangefilter==0.9.0
django-admin-tools==0.9.2
django-allauth==0.52.0
django-bulk-update==2.2.0
django-celery-results==2.4.0
django-chunked-iterator==0.6.1
django-cors-headers==3.8.0
django-countries==7.2.1
django-environ==0.9.0
django-filter==2.4.0
django-import-export==2.5.0
django-ipware==4.0.0
django-maintenance-mode==0.16.0
django-model-utils==4.1.1
django-modeltranslation==0.17.3
django-otp==1.0.6
django-pkgconf==0.4.0
django-qsstats-magic==1.1.0
django-redis==5.0.0
django-rest-swagger==2.2.0
django-tabbed-admin==1.0.4
django-user-agents==0.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
dnspython==2.1.0
docker-py==1.10.6
docker-pycreds==0.4.0
drf-spectacular==0.23.1
ecdsa==0.17.0
et-xmlfile==1.1.0
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
ethereum-input-decoder==0.2.2
executing==1.2.0
fabric==2.6.0
filelock==3.0.12
flake8==3.9.2
flower==1.2.0
fonttools==4.38.0
frozenlist==1.3.3
funcsigs==1.0.2
funcy==1.16
future==0.18.3
fuzzywuzzy==0.18.0
graypy==2.1.0
gunicorn==20.1.0
h11==0.12.0
hdwallets==0.1.2
hexbytes==0.2.2
hiredis==2.0.0
html5lib==1.1
httpcore==0.15.0
httpx==0.23.3
humanize==3.11.0
hyperlink==21.0.0
identify==2.2.13
idna==3.2
idna-ssl==1.1.0
immutables==0.16
importlib-metadata==4.8.1
importlib-resources==5.2.2
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
invoke==1.6.0
ipfshttpclient==0.8.0a2
ipython==8.10.0
ipython-genutils==0.2.0
isodate==0.6.0
itypes==1.2.0
janus==1.0.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.1
jsonfield==3.1.0
jsonrpcclient==4.0.0
jsonrpclib-pelix==0.4.2
jsonrpcserver==5.0.3
jsonschema==3.2.0
keyring==23.1.0
keyrings.alt==4.1.0
kombu==5.2.4
lru-dict==1.1.7
lxml==4.9.2
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mercadopago==2.2.0
mnemonic==0.20
model-mommy==2.0.0
more-itertools==8.8.0
msgpack==1.0.2
multiaddr==0.0.9
multidict==5.1.0
Naked==0.1.31
netaddr==0.8.0
nodeenv==1.6.0
numpy==1.24.2
oauthlib==3.2.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.7
OSlash==0.6.3
packaging==23.0
paramiko==2.7.2
parse==1.19.0
parse-type==0.5.2
parsimonious==0.8.1
parso==0.8.2
path==16.2.0
path.py==12.5.0
pathlib2==2.3.6
pbkdf2==1.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.3.0
pluggy==1.0.0
pre-commit==2.15.0
prettytable==2.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.37
protobuf==3.20.3
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.3.0
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.17
debugpy>=1.5.1
pydyf==0.5.0
pyflakes==2.3.1
Pygments==2.10.0
PyHamcrest==2.0.2
PyJWT==2.6.0
PyNaCl==1.4.0
pyOpenSSL==23.0.0
pyotp==2.6.0
pyparsing==2.4.7
pyphen==0.13.0
pyrsistent==0.18.0
pysha3==1.0.2
PySocks==1.7.1
py-solc-x==1.1.1
pytest==6.2.5
pytest-cov==2.12.1
pytest-django==4.4.0
python-binance>=1.0.12
python-bitcoinrpc==1.0
python-dateutil==2.8.2
python-Levenshtein==0.12.2
python-memcached==1.59
python-ripple-lib==1.0.11
python-socks==1.2.4
python-telegram-bot==13.15
python3-openid==3.2.0
pytz==2022.7.1
pywallet @ https://github.com/polyxexchange/pywallet/archive/master.zip
pyxdg==0.27
PyYAML==5.4.1
QDarkStyle==3.0.2
qrcode==7.3
QtPy==1.10.0
redis==3.5.3
regex==2021.8.28
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rlp==2.0.1
SecretStorage==3.3.1
sentry-sdk==1.3.1
service-identity==21.1.0
sha256>=0.1
shellescape==3.8.1
simplegeneric==0.8.1
simplejson==3.17.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.3
stack-data==0.6.2
tablib==3.0.0
tabulate==0.8.9
tblib==1.7.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.1.0
tronpy @ https://github.com/polyxexchange/tronpy/archive/refs/tags/0.2.10.zip
twilio==6.50.1
Twisted==22.10.0
txaio==21.2.1
typing_extensions==4.5.0
tzdata==2022.2
tzlocal==3.0
ua-parser==0.10.0
ujson==5.7.0
uritemplate==3.0.1
urllib3==1.26.6
user-agents==2.2.0
varint==1.0.2
vine==5.0.0
virtualenv==20.7.2
wcwidth==0.2.5
web3==5.23.1
webencodings==0.5.1
websocket-client==1.2.1
websockets==9.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.6.3
zeep==4.1.0
zipp==3.5.0
zope.interface==5.4.0
zopfli==0.2.1
